factor out store/cache to a new storage module
